requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
diff-match-patch==20121119
Django==1.9.8
django-crispy-forms==1.6.1
django-formtools==2.1
django-import-export==0.5.1
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
jdcal==1.3
MySQL-python==1.2.5
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
Pillow==4.3.0
PyYAML==3.12
requests==2.18.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
